Move code from UPnP2.java plugin file to new classes
